Rename eEURO -> EUROe
